# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/automation_rules:
    get:
      tags:
        - AutomationRules
      summary: List AutomationRules
      description: Get a paginated list of AutomationRules with optional filtering and sorting
      operationId: listAutomationRules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRuleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AutomationRules
      summary: Create AutomationRule
      description: Create a new AutomationRule
      operationId: createAutomationRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAutomationRuleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/automation_runs:
    get:
      tags:
        - AutomationRuns
      summary: List AutomationRuns
      description: Get a paginated list of AutomationRuns with optional filtering and sorting
      operationId: listAutomationRuns
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRunList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AutomationRuns
      summary: Create AutomationRun
      description: Create a new AutomationRun
      operationId: createAutomationRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAutomationRunRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRun'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    AutomationRule:
      type: object
      required:
        - id
        - name
        - model
        - trigger_kind
        - actions
        - active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        description:
          nullable: true
          type: string
          maxLength: 500
        model:
          type: string
          maxLength: 120
        trigger_kind:
          type: string
        trigger_pattern:
          nullable: true
          type: string
          maxLength: 160
        time_field:
          nullable: true
          type: string
          maxLength: 63
        delay_minutes:
          nullable: true
          type: integer
          format: int32
        actions:
          type: object
        active:
          type: boolean
        created_by_source:
          nullable: true
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreateAutomationRuleRequest:
      type: object
      required:
        - name
        - model
        - trigger_kind
        - actions
        - active
        - metadata
      properties:
        name:
          type: string
          maxLength: 120
        description:
          nullable: true
          type: string
          maxLength: 500
        model:
          type: string
          maxLength: 120
        trigger_kind:
          type: string
        trigger_pattern:
          nullable: true
          type: string
          maxLength: 160
        time_field:
          nullable: true
          type: string
          maxLength: 63
        delay_minutes:
          nullable: true
          type: integer
          format: int32
        actions:
          type: object
        active:
          type: boolean
        created_by_source:
          nullable: true
          type: string
          maxLength: 120
        metadata:
          type: object
...
    AutomationRuleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AutomationRule'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AutomationRun:
      type: object
      required:
        - id
        - automation_id
        - trigger
        - status
        - envelope_id
        - depth
        - actions_applied
        - effect_event_ids
        - occurred_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        automation_id:
          type: string
          format: uuid
        trigger:
          type: string
        status:
          type: string
        envelope_id:
          type: string
          maxLength: 64
        source_event_type:
          nullable: true
          type: string
          maxLength: 160
        aggregate_id:
          nullable: true
          type: string
          maxLength: 160
        parent_run_id:
          nullable: true
          type: string
          format: uuid
        depth:
          type: integer
          format: int32
        actions_applied:
          type: object
        effect_event_ids:
          type: object
        detail:
          nullable: true
          type: string
          maxLength: 500
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateAutomationRunRequest:
      type: object
      required:
        - automation_id
        - trigger
        - status
        - envelope_id
        - depth
        - actions_applied
        - effect_event_ids
        - occurred_at
        - metadata
      properties:
        automation_id:
          type: string
          format: uuid
        trigger:
          type: string
        status:
          type: string
        envelope_id:
          type: string
          maxLength: 64
        source_event_type:
          nullable: true
          type: string
          maxLength: 160
        aggregate_id:
          nullable: true
          type: string
          maxLength: 160
        parent_run_id:
          nullable: true
          type: string
          format: uuid
        depth:
          type: integer
          format: int32
        actions_applied:
          type: object
        effect_event_ids:
          type: object
        detail:
          nullable: true
          type: string
          maxLength: 500
        occurred_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    AutomationRunList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AutomationRun'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []